feat: do the UI of the different components
